refactor: 프로필 관련 스토리 디렉토리 이동
