+++ src/stories/pages/profile/ProfilePage.stories.tsx
@@ -1,16 +1,14 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import NavigationGroup from '~/components/NavigationGroup';
 import ProfilePage from '~/pages/profile/[id]';
 import { PageLayout } from '~/stories/Layout';
 
 const meta: Meta<typeof ProfilePage> = {
-  title: 'Page/Profile',
+  title: 'Page/Profile/Root',
   component: ProfilePage,
   decorators: [
     (Story) => (
       <PageLayout>
-        <NavigationGroup />
         <Story />
       </PageLayout>
     ),
@@ -22,6 +20,6 @@ const meta: Meta<typeof ProfilePage> = {
 
 export default meta;
 
-type SignInPageStory = StoryObj<typeof ProfilePage>;
+type ProfilePageStory = StoryObj<typeof ProfilePage>;
 
-export const Default: SignInPageStory = {};
+export const Default: ProfilePageStory = {};
